feat(server): Time control. Part 2. Application and api layer implementation (#197)
* feat(server): application remaining time logic. core logic and tests adjustment


* feat(server): time control. api and application layer implementation


* fix(client): generate new schema


* fix(server): test name


---------

+++ GomokuServer/src/GomokurServer.Core/Games/Entities/Clock.cs
@@ -5,7 +5,7 @@ namespace GomokuServer.Core.Games.Entities;
 public class Clock
 {
 	private long? _lastStartTime;
-	private long _remainingTimeInSeconds;
+	private int _remainingTimeInSeconds;
 	private readonly TimeControl _timeControl;
 	private readonly IDateTimeProvider _dateTimeProvider;
 
@@ -16,9 +16,9 @@ public class Clock
 		_dateTimeProvider = dateTimeProvider;
 	}
 
-	public long RemainingTimeInSeconds =>
+	public int RemainingTimeInSeconds =>
 		_lastStartTime.HasValue
-			? _remainingTimeInSeconds - (_dateTimeProvider.UtcNowInPosix - _lastStartTime.Value)
+			? _remainingTimeInSeconds - (int)(_dateTimeProvider.UtcNowInPosix - _lastStartTime.Value)
 			: _remainingTimeInSeconds;
 
 	public void Start()
@@ -33,13 +33,13 @@ public class Clock
 	{
 		if (_lastStartTime.HasValue)
 		{
-			_remainingTimeInSeconds -= (_dateTimeProvider.UtcNowInPosix - _lastStartTime.Value);
+			_remainingTimeInSeconds -= (int)(_dateTimeProvider.UtcNowInPosix - _lastStartTime.Value);
 			_remainingTimeInSeconds += _timeControl.IncrementPerMove;
 			_lastStartTime = null;
 		}
 	}
 
-	public void AddTime(long incrementInSeconds)
+	public void AddTime(int incrementInSeconds)
 	{
 		_remainingTimeInSeconds += incrementInSeconds;
 	}
